k8s-serverless/src/kubeless/moveToNextQueue/requirements.txt
kafka-python==1.3.5
requests==2.18.4